requirements.txt
fastapi
uvicorn[standard]
pydantic
networkx
websockets
openai
tenacity
openenv-core
